+++ BaseModule/Tasks/HeatTreatmentModule/DiffusionControl.cs
@@ -139,7 +139,7 @@ namespace BaseModule.Tasks.HeatTreatmentModule
             player.StopValue = (int)checkStopTime;
         }
 
-        private void player_CheckingEvent(object arg1, float arg2)
+        private void player_CheckingEvent(object arg1, int arg2)
         {
             CheckDataEvent(this, new CheckDataEventArgs(DataName, arg2));
         }
