+++ BaseModule/Tasks/HeatTreatmentModule/HeatControl.cs
@@ -206,7 +206,7 @@ namespace BaseModule.Tasks.HeatTreatmentModule
 
         }
 
-        private void player_CheckingEvent(object arg1, float arg2)
+        private void player_CheckingEvent(object arg1, int arg2)
         {
             CheckDataEvent(this, new CheckDataEventArgs(DataName, arg2));
         }
